backend/requirements.txt
fastapi
uvicorn
pandas
numpy
prophet
scikit-learn
groq
google-generativeai
python-dotenv
python-multipart